Extract common code from ComponentProcessingStep and ProductionComponentProcessingStep.
While working to fix a separate bug related to dagger/autofactory interaction, I noticed that these two steps follow the same pattern: For each "component" type, if validation passes, then create a BindingGraph and generate a class from it. Only the component type and what it meant for that to be valid are different.

This CL extracts the common stuff to a new base class. This will mean that the functional change I want to introduce to both can be done once instead of twice.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97232936